feature: generalize implementation of supplier and manufacturer classes
