+++ src/main/java/br/com/talles/domain/Manufacturer.java
@@ -3,43 +3,6 @@ package br.com.talles.domain;
 /**
  * Manufacturer entity class
  */
-public class Manufacturer extends Entity {
-    private String name;
-    private String registry;
-    private String phone;
-    private String email;
+public class Manufacturer extends Company {
 
-    public Manufacturer() {}
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getRegistry() {
-        return registry;
-    }
-
-    public void setRegistry(String registry) {
-        this.registry = registry;
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
 }
